requirements.txt
polars>=0.20.0
pyyaml>=6.0.0
